requirements.txt
opencv-python~=4.7.0.72
imread-from-url
onnxruntime~=1.16.0
cap-from-youtube
ultralytics==8.0.42
uvicorn[standard]==0.20.0
gunicorn==20.1.0
fastapi[all]~=0.100.0
loguru~=0.6.0
pytest
pytest-cov
numpy~=1.24.3
Pillow~=9.5.0
pydantic~=1.10.8